update wasm nodes

+++ wasm/rust/weather/src/lib.rs
@@ -1,7 +1,7 @@
 use space_lib::Request;
 
 #[no_mangle]
-fn main(lat: f32, lon: f32) -> Box<String> {
+fn main(lat: f64, lon: f64) -> Box<String> {
     let response = Request::get("https://api.openweathermap.org/data/2.5/weather")
         .query("lat", lat)
         .query("lon", lon)
